Issues #44 refactor: 중복 패키지 정리
