+++ presentation/src/main/java/com/lighthouse/presentation/extension/KotlinExtention.kt
@@ -1,4 +1,4 @@
-package com.lighthouse.presentation.utils.extention
+package com.lighthouse.presentation.extension
 
 import android.content.res.Resources
 import android.util.TypedValue
@@ -7,3 +7,9 @@ val Int.dp
     get() = Resources.getSystem().displayMetrics?.let { dm ->
         TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this.toFloat(), dm)
     } ?: 0f
+
+val screenWidth: Int
+    get() = Resources.getSystem().displayMetrics?.widthPixels ?: 0
+
+val screenHeight: Int
+    get() = Resources.getSystem().displayMetrics?.heightPixels ?: 0
