+++ presentation/src/main/java/com/lighthouse/presentation/extension/View.kt
@@ -5,8 +5,6 @@ import android.view.View
 fun View.isOnScreen(): Boolean {
     val viewLocation = IntArray(2)
     getLocationOnScreen(viewLocation)
-    val screenHeight = resources.displayMetrics.heightPixels
-    val screenWidth = resources.displayMetrics.widthPixels
 
     return viewLocation[0] in 0..screenWidth && viewLocation[1] in 0..screenHeight
 }
